requirements.txt
accelerate==1.12.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bitsandbytes==0.49.1
bleach==6.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
config==0.5.1
contourpy==1.3.3
cryptography==42.0.5
cycler==0.12.1
datasets==4.5.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
eval-type-backport==0.1.3
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.8.0.post1
fastjsonschema==2.19.1
filelock==3.25.2
fire==0.6.0
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
hf-xet==1.2.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface_hub==1.4.1
idna==3.6
igraph==0.11.6
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
ipywidgets==8.1.2
isoduration==20.11.0
jaxtyping==0.2.34
jedi==0.19.1
Jinja2==3.1.6
jiter==0.5.0
joblib==1.4.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kiwisolver==1.4.9
lightning-utilities==0.11.6
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
modelscope==1.34.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
openai==1.43.0
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
peft==0.10.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.0
plotly==5.24.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==23.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyDrive2==1.19.0
Pygments==2.17.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==6.9.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
regex==2023.12.25
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.7.0
scikit-learn==1.5.1
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.2
sentence-transformers==5.2.2
sentencepiece==0.2.0
sentry-sdk==2.52.0
setproctitle==1.3.3
setuptools==70.2.0
shellingham==1.5.4
shtab==1.7.1
simple_parsing==0.1.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.4.0
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.5.0
tinycss2==1.2.1
together==1.2.11
tokenizers==0.22.2
tomli==2.0.1
torchmetrics==1.4.1
tornado==6.4
tqdm==4.67.3
traitlets==5.14.2
transformers==5.1.0
triton==3.1.0
trl==0.13.0
typeguard==2.13.3
typer==0.12.5
typer-slim==0.21.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.15.0
tyro==0.8.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
wandb==0.24.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wheel==0.41.2
widgetsnbextension==4.0.10
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
